Rename function. Refs #74
