+++ src/functions/index.js
@@ -12,7 +12,7 @@ const scheduledTask = require('../scheduler/scheduledTask')();
 const ipfsService = require('../services').ipfsService;
 const IpfsCatFunction = require('./IpfsCatFunction');
 const web3 = require('../globalWeb3').web3;
-const SendRawTxFunction = require('/SendRawTxFunction');
+const SendRawTxFunction = require('./SendRawTxFunction');
 
 handlers.push(new IpfsCatFunction(scheduledTask, ipfsService));
 handlers.push(new SendRawTxFunction(scheduledTask, web3));
